AccommodationAPI/accommodations/perms.py
from rest_framework import permissions

class OwnerPerms(permissions.IsAuthenticated):
    def has_object_permission(self, request, view, obj):
        return obj.owner == request.user and request.user == obj.user